requirements.txt
Django>=4.2,<5.0  # For REST API
djangorestframework==3.14.0  # For REST API
django-cors-headers==4.3.1   # For handling CORS
django-filter==23.5          # For advanced filtering
psycopg2-binary==2.9.9      # For PostgreSQL database
python-dotenv==1.0.0        # For environment variables
drf-spectacular==0.27.0     # For API documentation
Pillow==10.1.0             # For image handling
pandas==2.1.4              # For data analysis
numpy==1.26.3              # For numerical computations
celery==5.3.6             # For async tasks
redis==5.0.1              # For caching and Celery backend
django-celery-beat==2.5.0  # For periodic tasks
djangorestframework-simplejwt==5.3.1
matplotlib==3.8.2
seaborn==0.13.1
reportlab==4.0.9
scikit-learn==1.3.2
plotly==5.18.0
channels==4.0.0
channels-redis==4.1.0
